+++ src/Ext/Techno/Body.Update.cpp
@@ -7,6 +7,7 @@
 #include <Ext/Bullet/Body.h>
 #include <Ext/WarheadType/Body.h>
 #include <Ext/House/Body.h>
+#include <Ext/Script/Body.h>
 #include <Ext/WeaponType/Body.h>
 #include <Utilities/EnumFunctions.h>
 #include <Utilities/AresFunctions.h>
@@ -930,4 +931,50 @@ void TechnoExt::ExtData::UpdateDelayFire()
 
 	if (delayedFire_Duration > 0)
 		this->DelayedFire_DurationTimer.Start(delayedFire_Duration);
+}
+
+void TechnoExt::ExtData::RefreshRandomTargets()
+{
+	auto const pThis = this->OwnerObject();
+	if (!pThis)
+		return;
+
+	const auto pExt = TechnoExt::ExtMap.Find(pThis);
+	if (!pExt)
+		return;
+
+	if (pThis->Target && pThis->SpawnManager && pExt->CurrentRandomTarget && ScriptExt::IsUnitAvailable(static_cast<TechnoClass*>(pExt->CurrentRandomTarget), true))
+	{
+		if (pThis->SpawnManager)
+		{
+			for (auto pSpawn : pThis->SpawnManager->SpawnedNodes)
+			{
+				if (!pSpawn->Unit)
+					continue;
+
+				auto pSpawnExt = TechnoExt::ExtMap.Find(pSpawn->Unit);
+				if (!pSpawnExt)
+					continue;
+
+				if (!pSpawnExt->CurrentRandomTarget)
+				{
+					pSpawnExt->CurrentRandomTarget = TechnoExt::GetRandomTarget(pThis);
+					pSpawn->Unit->Target = pSpawnExt->CurrentRandomTarget;
+				}
+				else if (pSpawn->Status == SpawnNodeStatus::Preparing && pSpawn->Unit->IsInAir())
+				{
+					if (!pSpawn->Unit->Target && pSpawnExt->CurrentRandomTarget)
+						pSpawn->Unit->Target = pSpawnExt->CurrentRandomTarget;
+				}
+			}
+		}
+	}
+
+	if (pExt->OriginalTarget && !pThis->Target && ScriptExt::IsUnitAvailable(static_cast<TechnoClass*>(pExt->OriginalTarget), true) && !pThis->IsInAir())
+	{
+		if (pExt->CurrentRandomTarget && ScriptExt::IsUnitAvailable(pExt->CurrentRandomTarget, true))
+			pThis->SetTarget(pExt->CurrentRandomTarget);
+		else
+			pThis->SetTarget(pExt->OriginalTarget);
+	}
 }