+++ src/Ext/Techno/Body.cpp
@@ -7,6 +7,8 @@
 #include <TacticalClass.h>
 
 #include <Ext/House/Body.h>
+#include <Ext/WeaponType/Body.h>
+#include <Ext/Script/Body.h>
 #include <Ext/Team/Body.h>
 #include <Ext/Script/Body.h>
 
@@ -639,6 +641,213 @@ bool TechnoExt::IsTypeImmune(TechnoClass* pThis, TechnoClass* pSource)
 	return false;
 }
 
+bool TechnoExt::UpdateRandomTarget(TechnoClass* pThis)
+{
+	if (!pThis)
+		return false;
+
+	int weaponIndex = pThis->SelectWeapon(pThis->Target);
+	auto pWeapon = pThis->GetWeapon(weaponIndex)->WeaponType;
+	if (!pWeapon)
+		return false;
+
+	const auto pWeaponExt = WeaponTypeExt::ExtMap.Find(pWeapon);
+	if (!pWeaponExt || pWeaponExt->RandomTarget <= 0.0)
+		return false;
+
+	const auto pExt = TechnoExt::ExtMap.Find(pThis);
+	if (!pExt)
+		return false;
+
+	if (pThis->GetCurrentMission() == Mission::Move)
+	{
+		pExt->CurrentRandomTarget = nullptr;
+		pExt->OriginalTarget = nullptr;
+
+		return false;
+	}
+
+	if (pExt->CurrentRandomTarget && ScriptExt::IsUnitAvailable(pExt->CurrentRandomTarget, false) && pThis->SpawnManager)
+		return false;
+
+	if (!pThis->Target && !ScriptExt::IsUnitAvailable(abstract_cast<TechnoClass*>(pExt->OriginalTarget), false))
+	{
+		pExt->OriginalTarget = nullptr;
+		return false;
+	}
+
+	if (pExt->OriginalTarget && !ScriptExt::IsUnitAvailable(abstract_cast<TechnoClass*>(pExt->OriginalTarget), false))
+	{
+		pExt->CurrentRandomTarget = nullptr;
+		pExt->OriginalTarget = nullptr;
+	}
+
+	if (pThis->GetCurrentMission() != Mission::Attack)
+	{
+		pExt->OriginalTarget = nullptr;
+		return false;
+	}
+
+	if (!pThis->Target)
+		return false;
+
+	if (pThis->DistanceFrom(pExt->OriginalTarget) > pWeapon->Range)
+	{
+		if (pThis->WhatAmI() == AbstractType::Building)
+		{
+			pThis->SetTarget(nullptr);
+			pExt->CurrentRandomTarget = nullptr;
+			pExt->OriginalTarget = nullptr;
+
+			return false;
+		}
+
+		pThis->SetTarget(pExt->OriginalTarget);
+	}
+
+	if (pThis->DistanceFrom(pThis->Target) > pWeapon->Range)
+	{
+		pThis->SetTarget(pExt->OriginalTarget);
+		return false;
+	}
+
+	auto pRandomTarget = GetRandomTarget(pThis);
+
+	if (!pRandomTarget)
+		return false;
+
+	pExt->OriginalTarget = !pExt->OriginalTarget ? pThis->Target : pExt->OriginalTarget;
+	pExt->CurrentRandomTarget = pRandomTarget;
+	pThis->Target = pRandomTarget;
+
+	if (pThis->SpawnManager)
+	{
+		bool isFirstSpawn = true;
+
+		for (auto pSpawn : pThis->SpawnManager->SpawnedNodes)
+		{
+			if (!pSpawn->Unit)
+				continue;
+
+			TechnoClass* pSpawnTarget = nullptr;
+
+			auto pSpawnExt = TechnoExt::ExtMap.Find(pSpawn->Unit);
+			if (!pSpawnExt)
+				continue;
+
+			if (isFirstSpawn)
+			{
+				pSpawnTarget = pExt->CurrentRandomTarget;
+
+				if (pWeaponExt->RandomTarget_Spawners_MultipleTargets)
+					isFirstSpawn = false;
+			}
+			else
+			{
+				pSpawnTarget = GetRandomTarget(pThis);
+
+				if (!pSpawnTarget)
+					pSpawnTarget = abstract_cast<TechnoClass*>(pExt->OriginalTarget);
+			}
+
+			pSpawnExt->CurrentRandomTarget = pSpawnTarget;
+			pSpawnExt->OriginalTarget = pExt->OriginalTarget;
+		}
+	}
+
+	return true;
+}
+
+TechnoClass* TechnoExt::GetRandomTarget(TechnoClass* pThis)
+{
+	TechnoClass* selection = nullptr;
+
+	if (!pThis && !pThis->Target)
+		return selection;
+
+	int weaponIndex = pThis->SelectWeapon(pThis->Target);
+	auto pWeapon = pThis->GetWeapon(weaponIndex)->WeaponType;
+	if (!pWeapon)
+		return selection;
+
+	const auto pWeaponExt = WeaponTypeExt::ExtMap.Find(pWeapon);
+	if (!pWeaponExt || pWeaponExt->RandomTarget <= 0.0)
+		return selection;
+
+	const auto pExt = TechnoExt::ExtMap.Find(pThis);
+	if (!pExt)
+		return selection;
+
+	int retargetProbability = std::min((int)round(pWeaponExt->RandomTarget * 100), 100);
+	int dice = ScenarioClass::Instance->Random.RandomRanged(1, 100);
+
+	if (retargetProbability < dice)
+		return selection;
+
+	auto pThisType = pThis->GetTechnoType();
+	int minimumRange = pWeapon->MinimumRange;
+	int range = pWeapon->Range;
+	int airRange = pWeapon->Range + pThisType->AirRangeBonus;
+	bool omniFire = pWeapon->OmniFire;
+	std::vector<TechnoClass*> candidates;
+	auto originalTarget = abstract_cast<TechnoClass*>(!pExt->OriginalTarget ? pThis->Target : pExt->OriginalTarget);
+	bool friendlyFire = pThis->Owner->IsAlliedWith(originalTarget);
+
+	// Looking for all valid targeting candidates
+	for (auto pCandidate : *TechnoClass::Array)
+	{
+		if (pCandidate == pThis
+			|| !ScriptExt::IsUnitAvailable(pCandidate, true)
+			|| pThisType->Immune
+			|| !EnumFunctions::IsTechnoEligible(pCandidate, pWeaponExt->CanTarget, true)
+			|| (!pWeapon->Projectile->AA && pCandidate->IsInAir())
+			|| (!pWeapon->Projectile->AG && !pCandidate->IsInAir())
+			|| (!friendlyFire && (pThis->Owner->IsAlliedWith(pCandidate) || ScriptExt::IsUnitMindControlledFriendly(pThis->Owner, pCandidate)))
+			|| pCandidate->TemporalTargetingMe
+			|| pCandidate->BeingWarpedOut
+			|| (pCandidate->GetTechnoType()->Underwater && pCandidate->GetTechnoType()->NavalTargeting == NavalTargetingType::Underwater_Never)
+			|| (pCandidate->GetTechnoType()->Naval && pCandidate->GetTechnoType()->NavalTargeting == NavalTargetingType::Naval_None)
+			|| (pCandidate->CloakState == CloakState::Cloaked && !pThisType->Naval)
+			|| (pCandidate->InWhichLayer() == Layer::Underground))
+		{
+			continue;
+		}
+
+		int distanceFromAttacker = pThis->DistanceFrom(pCandidate);
+		if (distanceFromAttacker < minimumRange)
+			continue;
+
+		if (omniFire)
+		{
+			if (pCandidate->IsInAir())
+				range = airRange;
+
+			if (distanceFromAttacker <= range)
+				candidates.push_back(pCandidate);
+		}
+		else
+		{
+			int distanceFromOriginalTargetToCandidate = pCandidate->DistanceFrom(pThis->Target);
+			int distanceFromOriginalTarget = pThis->DistanceFrom(pThis->Target);
+
+			if (pCandidate->IsInAir())
+				range = airRange;
+
+			if (distanceFromAttacker <= range && distanceFromOriginalTargetToCandidate <= distanceFromOriginalTarget)
+				candidates.push_back(pCandidate);
+		}
+	}
+
+	if (candidates.size() == 0)
+		return selection;
+
+	// Pick one new target from the list of targets inside the weapon range
+	dice = ScenarioClass::Instance->Random.RandomRanged(0, candidates.size() - 1);
+	selection = candidates.at(dice);
+
+	return selection;
+}
+
 // =============================
 // load / save
 
@@ -661,6 +870,8 @@ void TechnoExt::ExtData::Serialize(T& Stm)
 		.Process(this->DeployFireTimer)
 		.Process(this->ForceFullRearmDelay)
 		.Process(this->WHAnimRemainingCreationInterval)
+		.Process(this->OriginalTarget)
+		.Process(this->CurrentRandomTarget)
 		.Process(this->DelayedFire_Charging)
 		.Process(this->DelayedFire_Charged)
 		.Process(this->DelayedFire_Anim)