+++ src/Ext/WeaponType/Body.cpp
@@ -59,6 +59,8 @@ void WeaponTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->OmniFire_TurnToTarget.Read(exINI, pSection, "OmniFire.TurnToTarget");
 	this->ExtraWarheads.Read(exINI, pSection, "ExtraWarheads");
 	this->ExtraWarheads_DamageOverrides.Read(exINI, pSection, "ExtraWarheads.DamageOverrides");
+	this->RandomTarget.Read(exINI, pSection, "RandomTarget");
+	this->RandomTarget_Spawners_MultipleTargets.Read(exINI, pSection, "RandomTarget.Spawners.MultipleTargets");
 	this->DelayedFire_Anim.Read(exINI, pSection, "DelayedFire.Anim");
 	this->DelayedFire_PostAnim.Read(exINI, pSection, "DelayedFire.PostAnim");
 	this->DelayedFire_Anim_UseFLH.Read(exINI, pSection, "DelayedFire.Anim.UseFLH");
@@ -89,6 +91,8 @@ void WeaponTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->OmniFire_TurnToTarget)
 		.Process(this->ExtraWarheads)
 		.Process(this->ExtraWarheads_DamageOverrides)
+		.Process(this->RandomTarget)
+		.Process(this->RandomTarget_Spawners_MultipleTargets)
 		.Process(this->DelayedFire_Anim)
 		.Process(this->DelayedFire_PostAnim)
 		.Process(this->DelayedFire_Anim_UseFLH)