+++ docs/Whats-New.md
@@ -350,6 +350,7 @@ New:
 - Allow enabling application of `Verses` and `PercentAtMax` for negative damage (by Starkku)
 - Delayed fire animation (by FS-21)
 - Script actions for manipulating script flows with conditional jumps (by FS-21)
+- `Burst.Retarget` for assigning a new target in each projectile (by FS-21)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)